Fully resolve file path imports
This allows us to get the same behavior of this plugin for local
modules as well. eg, we have a local directory with lots of
components, and, for convenience, an index.js which just re-exports
all of the components.

Other local code can refer to these components by eg `import {A} from
'./components';`, so it'd be useful to use this plugin's
behavior. However, because the import source is a file path rather
than a library name, we currently cannot setup the transform
options (since the source could be './components', or '../components',
or '../../components', etc.). This PR solves that situation by fully
resolving the file paths.